+++ app/Auth/index.js
@@ -1,0 +1,4 @@
+/**
+ * Main Auth Router
+ */
+module.exports = require('./routes')