+++ database/seeds/user.js
@@ -1,18 +1,9 @@
-const bcrypt = require('bcrypt')
-
 const user = {
   name: 'Jhon Doe',
-  email: 'admin@admin.com',
-  password: '123456'
+  email: 'admin@admin.com'
 }
 
 exports.seed = function (knex, Promise) {
   return knex('users').del()
-    .then(() =>
-      bcrypt.hash(user.password, 10)
-        .then((hash) => {
-          user.password = hash
-          return knex('users').insert([user])
-        })
-    )
+    .then(() => knex('users').insert([user]))
 }